add: reviewerId validation on delete review

+++ src/domain/review/application/services/delete-review.ts
@@ -1,10 +1,11 @@
 import { Either, left, right } from '@/core/either'
-import { Injectable } from '@nestjs/common'
+import { Injectable, UnauthorizedException } from '@nestjs/common'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
 import { ReviewsRepository } from '../repositories/reviews-repository'
 
 interface DeleteReviewServiceRequest {
   reviewId: string
+  reviewerId: string
 }
 
 type DeleteReviewServiceResponse = Either<ResourceNotFoundError, null>
@@ -15,11 +16,15 @@ export class DeleteReviewService {
 
   async execute({
     reviewId,
+    reviewerId,
   }: DeleteReviewServiceRequest): Promise<DeleteReviewServiceResponse> {
     const reviewToDelete = await this.reviewsRepository.findById(reviewId)
 
     if (!reviewToDelete) {
       return left(new ResourceNotFoundError())
+    }
+    if (reviewToDelete.reviewerId.toString() !== reviewerId) {
+      return left(new UnauthorizedException())
     }
 
     await this.reviewsRepository.delete(reviewToDelete)
